Overflow fix

+++ src/c8device.rs
@@ -99,7 +99,7 @@ impl C8Device {
             const SHIFT: u8 = 8;
 
             let pc = self.program_counter as usize;
-            let opcode = ((self.memory[pc] << SHIFT) | self.memory[pc + 1]) as u16;
+            let opcode = (self.memory[pc] as u16) << SHIFT | self.memory[pc + 1] as u16;
 
             self.program_counter += 2;
